src/agents/functions/requirements.txt
firebase_functions~=0.1.0

flask
python-dotenv
firebase-admin
google-generativeai
